+++ framework/model/src/main/java/is/codion/framework/model/EntityQueryModel.java
@@ -18,6 +18,8 @@
  */
 package is.codion.framework.model;
 
+import is.codion.common.Conjunction;
+import is.codion.common.observer.Mutable;
 import is.codion.common.state.State;
 import is.codion.common.state.StateObserver;
 import is.codion.common.value.Value;
@@ -28,6 +30,7 @@ import is.codion.framework.domain.entity.EntityDefinition;
 import is.codion.framework.domain.entity.EntityType;
 import is.codion.framework.domain.entity.OrderBy;
 import is.codion.framework.domain.entity.attribute.Attribute;
+import is.codion.framework.domain.entity.condition.Condition;
 
 import java.util.List;
 import java.util.function.Function;
@@ -78,6 +81,26 @@ public interface EntityQueryModel extends Supplier<List<Entity>> {
 	 * @return the {@link EntityConditions} instance used by this query model
 	 */
 	EntityConditions conditions();
+
+	/**
+	 * Controls the additional WHERE condition, which can be used in conjunction with {@link #conditions()}.
+	 * The condition supplier may return null in case of no condition.
+	 * Note that in order for the {@link #conditionChanged()} {@link StateObserver} to indicate
+	 * a changed condition, the additional condition must be set via {@link AdditionalCondition#set(Object)},
+	 * changing the return value of the underlying {@link Supplier} instance does not trigger a changed condition.
+	 * @return the {@link AdditionalCondition} instance controlling the additional WHERE condition
+	 */
+	AdditionalCondition where();
+
+	/**
+	 * Controls the additional HAVING condition, which can be used in conjunction with {@link #conditions()}.
+	 * The condition supplier may return null in case of no condition.
+	 * Note that in order for the {@link #conditionChanged()} {@link StateObserver} to indicate
+	 * a changed condition, the additional condition must be set via {@link AdditionalCondition#set(Object)},
+	 * changing the return value of the underlying {@link Supplier} instance does not trigger a changed condition.
+	 * @return the {@link AdditionalCondition} instance controlling the additional HAVING condition
+	 */
+	AdditionalCondition having();
 
 	/**
 	 * Returns a {@link State} controlling whether this query model should query all underlying entities
@@ -145,5 +168,17 @@ public interface EntityQueryModel extends Supplier<List<Entity>> {
 	 */
 	static EntityQueryModel entityQueryModel(EntityConditions entityConditions) {
 		return new DefaultEntityQueryModel(entityConditions);
+	}
+
+	/**
+	 * Specifies an additional condition supplier.
+	 */
+	interface AdditionalCondition extends Mutable<Supplier<Condition>> {
+
+		/**
+		 * Default {@link Conjunction#AND}.
+		 * @return the {@link Mutable} controlling the {@link Conjunction} to use when adding the additional condition
+		 */
+		Mutable<Conjunction> conjunction();
 	}
 }