+++ framework/model/src/test/java/is/codion/framework/model/DefaultEntityConditionsTest.java
@@ -18,15 +18,12 @@
  */
 package is.codion.framework.model;
 
-import is.codion.common.Conjunction;
 import is.codion.common.db.exception.DatabaseException;
 import is.codion.common.model.condition.ConditionModel;
 import is.codion.common.user.User;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.db.local.LocalEntityConnectionProvider;
 import is.codion.framework.domain.entity.Entity;
-import is.codion.framework.domain.entity.condition.Condition;
-import is.codion.framework.domain.entity.condition.Condition.Combination;
 import is.codion.framework.model.test.TestDomain;
 import is.codion.framework.model.test.TestDomain.Department;
 import is.codion.framework.model.test.TestDomain.Detail;
@@ -128,27 +125,5 @@ public class DefaultEntityConditionsTest {
 		assertTrue(conditionModel.get(Employee.DEPARTMENT_FK).enabled().get());
 		conditionModel.clear();
 		assertFalse(conditionModel.get(Employee.DEPARTMENT_FK).enabled().get());
-	}
-
-	@Test
-	void condition() throws DatabaseException {
-		Entity sales = CONNECTION_PROVIDER.connection().selectSingle(Department.NAME.equalTo("SALES"));
-		Entity accounting = CONNECTION_PROVIDER.connection().selectSingle(Department.NAME.equalTo("ACCOUNTING"));
-		conditionModel.clear();
-		assertFalse(conditionModel.get(Employee.DEPARTMENT_FK).enabled().get());
-		conditionModel.setInOperands(Employee.DEPARTMENT_FK, asList(sales, accounting));
-		Condition condition = conditionModel.where(Conjunction.AND);
-		assertFalse(condition instanceof Combination);
-		conditionModel.additionalWhere().set(() -> Condition.custom(Employee.CONDITION_2_TYPE));
-		assertNotNull(conditionModel.additionalWhere().get());
-		condition = conditionModel.where(Conjunction.AND);
-		assertInstanceOf(Combination.class, condition);
-		assertEquals(Conjunction.AND, ((Combination) condition).conjunction());
-		conditionModel.additionalWhere().conjunction().set(Conjunction.OR);
-		condition = conditionModel.where(Conjunction.AND);
-		assertEquals(Conjunction.OR, ((Combination) condition).conjunction());
-		conditionModel.additionalWhere().set(null);
-		condition = conditionModel.where(Conjunction.AND);
-		assertFalse(condition instanceof Combination);
 	}
 }